pyproject.toml
[project]
name = "patient-advocacy-agent"
version = "0.1.0"
description = "patient_advocacy_agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "svlearn-bootcamp>=0.1.7",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "pillow>=9.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "faiss-cpu>=1.7.4",
    "pydantic>=2.0.0",
    "langchain>=0.1.0",
    "python-dotenv>=1.0.0",
    "sentencepiece>=0.1.99",
    "huggingface-hub>=0.17.0",
    "google-cloud-storage>=3.4.1",
    "anthropic>=0.39.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

# If you also build sdists and want data included:
[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    # ... other includes like README, pyproject.toml, etc.
]

[tool.hatch.build.targets.wheel]
packages = ["src/patient_advocacy_agent"]
